+++ udata_ckan/__init__.py
@@ -4,5 +4,5 @@ CKAN integration for udata
 '''
 from __future__ import unicode_literals
 
-__version__ = '1.0.2.dev'
+__version__ = '1.1.0.dev'
 __description__ = 'CKAN integration for udata'